+++ ChangeLog.md
@@ -1,8 +1,14 @@
 # Changelog for prairie
 
+## 0.0.3.0
+
+- [#8](https://github.com/parsonsmatt/prairie/pull/8)
+    - The `Prairie.Fold` module is introduced, allowing you to fold records.
+    - The `Prairie.Traverse` module is introduced, allowing you to traverse over records.
+
 ## 0.0.2.1
 
-- [#3](https://github.com/parsonsmatt/prairie/pull/6)
+- [#6](https://github.com/parsonsmatt/prairie/pull/6)
     - Bump upper bound for `TemplateHaskell`, supporting up to GHC 9.8
     - Fix string literals in docs that were quoted as module names